demo app/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
opencv-python
numpy
deepface
python-dotenv
groq
requests